Move BoundingBox
This places the definition and implementation of the class BoundingBox
into its own header and implementation file. This is done to cleanup
vds.cpp and as a preparation for future commits that will introduce
additional classes. These classes will also use separate implementation
and header files.
